Extract Specter.PeerConnection as separate module
