+++ lib/specter/native.ex
@@ -3,6 +3,7 @@ defmodule Specter.Native do
   use Rustler, otp_app: :specter, crate: :specter_nif
 
   @opaque t() :: reference()
+  @type peer_conn_t() :: Specter.PeerConnection.t()
 
   @doc """
   Initialize the NIF with RTC configuration, registering the current
@@ -18,33 +19,33 @@ defmodule Specter.Native do
   Asynchronously adds the candidate to the given RTCPeerConnection.
   Sends back `{:ok, _uuid, :add_ice_candidate}` when successful.
   """
-  @spec add_ice_candidate(t(), Specter.peer_connection_t(), Specter.ice_candidate_t()) ::
+  @spec add_ice_candidate(t(), peer_conn_t(), Specter.PeerConnection.ice_candidate_t()) ::
           :ok | {:error, term()}
   def add_ice_candidate(_ref, _pc, _candidate), do: error()
 
   @doc """
   Closes an RTCPeerConnection represented by the given UUID.
   """
-  @spec close_peer_connection(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec close_peer_connection(t(), peer_conn_t()) :: :ok | {:error, term()}
   def close_peer_connection(_ref, _pc), do: error()
 
   @doc """
   Create an answer from an RTCPeerConnection that has been given a remote description.
   """
-  @spec create_answer(t(), Specter.peer_connection_t(), boolean) :: :ok | {:error, term()}
+  @spec create_answer(t(), peer_conn_t(), boolean) :: :ok | {:error, term()}
   def create_answer(_ref, _pc, _vad), do: error()
 
   @doc """
   Add a data channel to an RTCPeerConnection.
   """
-  @spec create_data_channel(t(), Specter.peer_connection_t(), String.t()) ::
+  @spec create_data_channel(t(), peer_conn_t(), String.t()) ::
           :ok | {:error, term()}
   def create_data_channel(_ref, _pc, _label), do: error()
 
   @doc """
   Create an offer from an RTCPeerConnection.
   """
-  @spec create_offer(t(), Specter.peer_connection_t(), bool(), bool()) :: :ok | {:error, term()}
+  @spec create_offer(t(), peer_conn_t(), bool(), bool()) :: :ok | {:error, term()}
   def create_offer(_ref, _pc, _vad, _ice_restart), do: error()
 
   @doc """
@@ -58,39 +59,39 @@ defmodule Specter.Native do
   @doc """
   Sends back state of peer connection.
   """
-  @spec connection_state(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec connection_state(t(), peer_conn_t()) :: :ok | {:error, term()}
   def connection_state(_ref, _pc), do: error()
 
   @doc """
   Sends back the current session description. Will be nil until the peer connection successfully
   negotiates ICE, even if an offer or answer has been set as the local description.
   """
-  @spec current_local_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec current_local_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def current_local_description(_ref, _pc), do: error()
 
   @doc """
   Sends back the current remote session description. Will be nil until the peer connection successfully
   negotiates ICE, even if an offer or answer has been set as the remote description.
   """
-  @spec current_remote_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec current_remote_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def current_remote_description(_ref, _pc), do: error()
 
   @doc """
   Sends back state of ICE connection.
   """
-  @spec ice_connection_state(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec ice_connection_state(t(), peer_conn_t()) :: :ok | {:error, term()}
   def ice_connection_state(_ref, _pc), do: error()
 
   @doc """
   Sends back state of ICE gathering process.
   """
-  @spec ice_gathering_state(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec ice_gathering_state(t(), peer_conn_t()) :: :ok | {:error, term()}
   def ice_gathering_state(_ref, _pc), do: error()
 
   @doc """
   Sends back the pending or current session description, depending on the state of the connection.
   """
-  @spec local_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec local_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def local_description(_ref, _pc), do: error()
 
   @doc """
@@ -114,7 +115,7 @@ defmodule Specter.Native do
   - https://github.com/webrtc-rs/webrtc/blob/master/src/peer_connection/mod.rs
   """
   @spec new_peer_connection(t(), Specter.api_t()) ::
-          {:ok, Specter.peer_connection_t()} | {:error, term()}
+          {:ok, peer_conn_t()} | {:error, term()}
   def new_peer_connection(_ref, _api), do: error()
 
   @doc """
@@ -142,7 +143,7 @@ defmodule Specter.Native do
   Checks whether the UUID representing an RTCPeerConnection points to an initialized
   instance.
   """
-  @spec peer_connection_exists(t(), Specter.peer_connection_t()) ::
+  @spec peer_connection_exists(t(), peer_conn_t()) ::
           {:ok, boolean()} | {:error, term()}
   def peer_connection_exists(_ref, _pc), do: error()
 
@@ -150,14 +151,14 @@ defmodule Specter.Native do
   Sends back the pending session description. May be nil after the peer connection successfully
   negotiates its connection.
   """
-  @spec pending_local_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec pending_local_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def pending_local_description(_ref, _pc), do: error()
 
   @doc """
   Sends back the pending remote session description. May be nil after the peer connection successfully
   negotiates its connection.
   """
-  @spec pending_remote_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec pending_remote_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def pending_remote_description(_ref, _pc), do: error()
 
   @doc """
@@ -170,14 +171,14 @@ defmodule Specter.Native do
   @doc """
   Sends back the pending or current remote session description, depending on the state of the connection.
   """
-  @spec remote_description(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec remote_description(t(), peer_conn_t()) :: :ok | {:error, term()}
   def remote_description(_ref, _pc), do: error()
 
   @doc """
   Given a UUID representing an RTCPeerConnection and an offer or an answer from that same
   peer connection, set it as the local session description.
   """
-  @spec set_local_description(t(), Specter.peer_connection_t(), Specter.session_description_t()) ::
+  @spec set_local_description(t(), peer_conn_t(), Specter.PeerConnection.session_description_t()) ::
           :ok | {:error, term()}
   def set_local_description(_ref, _pc, _desc), do: error()
 
@@ -186,14 +187,14 @@ defmodule Specter.Native do
   answer from a different peer connection, set it on the peer connection as the remote session
   description.
   """
-  @spec set_remote_description(t(), Specter.peer_connection_t(), Specter.session_description_t()) ::
+  @spec set_remote_description(t(), peer_conn_t(), Specter.PeerConnection.session_description_t()) ::
           :ok | {:error, term()}
   def set_remote_description(_ref, _pc, _desc), do: error()
 
   @doc """
   Sends back state of sesion parameters negotiation.
   """
-  @spec signaling_state(t(), Specter.peer_connection_t()) :: :ok | {:error, term()}
+  @spec signaling_state(t(), peer_conn_t()) :: :ok | {:error, term()}
   def signaling_state(_ref, _pc), do: error()
 
   ##
