+++ test/support/nif_helpers.ex
@@ -3,13 +3,13 @@ defmodule SpecterTest.NifHelpers do
   import ExUnit.Assertions
 
   def create_data_channel(%{specter: specter, peer_connection: pc}) do
-    :ok = Specter.create_data_channel(specter, pc, "data")
+    :ok = Specter.PeerConnection.create_data_channel(specter, pc, "data")
     ExUnit.Assertions.assert_receive({:data_channel_created, ^pc})
     :ok
   end
 
   def create_offer(%{specter: specter, peer_connection: pc}) do
-    :ok = Specter.create_offer(specter, pc)
+    :ok = Specter.PeerConnection.create_offer(specter, pc)
     ExUnit.Assertions.assert_receive({:offer, ^pc, offer})
     [offer: offer]
   end
@@ -40,9 +40,9 @@ defmodule SpecterTest.NifHelpers do
   end
 
   def init_peer_connection(%Specter{} = specter, api) do
-    {:ok, pc} = Specter.new_peer_connection(specter, api)
+    {:ok, pc} = Specter.PeerConnection.new(specter, api)
     assert_receive {:peer_connection_ready, ^pc}
-    true = Specter.peer_connection_exists?(specter, pc)
+    true = Specter.PeerConnection.exists?(specter, pc)
     pc
   end
 end
